Simplify by improving CalendarModel object modification code

+++ core/Model/CalendarModel.vala
@@ -402,9 +402,9 @@ public class Maya.Model.CalendarModel : Object {
         });
     }
 
-    private void debug_event (E.Source source, ECal.Component event) {
+    private void debug_event (E.Source source, ECal.Component event, string message = "") {
         unowned ICal.Component comp = event.get_icalcomponent ();
-        debug (@"Event ['$(comp.get_summary())', $(source.dup_display_name()), $(comp.get_uid()))]");
+        debug (@"$(message) Event ['$(comp.get_summary())', $(source.dup_display_name()), $(comp.get_uid()), $(comp.get_dtstart().as_ical_string ())]");
     }
 
     //--- Signal Handlers ---//
@@ -434,7 +434,8 @@ public class Maya.Model.CalendarModel : Object {
 #else
             client.generate_instances_for_object_sync (comp, (time_t) data_range.first_dt.to_unix (), (time_t) data_range.last_dt.to_unix (), (event, start, end) => {
 #endif
-                debug_event (source, event);
+                debug_event (source, event, "ADDED");
+                event.set_data<E.Source>("source", source);
                 events.set (uid, event);
                 added_events.add (event);
                 return true;
@@ -451,25 +452,39 @@ public class Maya.Model.CalendarModel : Object {
 #endif
         debug (@"Received $(objects.length()) modified event(s) for source '%s'", source.dup_display_name ());
         var updated_events = new Gee.ArrayList<ECal.Component> ((Gee.EqualDataFunc<ECal.Component>?) Util.calcomponent_equal_func);
+        var removed_events = new Gee.ArrayList<ECal.Component> ((Gee.EqualDataFunc<ECal.Component>?) Util.calcomponent_equal_func);
+        var added_events = new Gee.ArrayList<ECal.Component> ((Gee.EqualDataFunc<ECal.Component>?) Util.calcomponent_equal_func);
         objects.foreach ((comp) => {
             unowned string uid = comp.get_uid ();
             var events_for_source = source_events.get (source);
             var events_for_uid = events_for_source.get (uid);
-            events_for_source.remove_all (uid);
+            if (events_for_uid.size > 1) {
+                events_for_source.remove_all (uid);
+                foreach (ECal.Component event in events_for_uid) {
+                    debug_event (source, event, "MODIFIED - ORIGINAL");
+                    removed_events.add (event);
+                }
 
 #if E_CAL_2_0
-            client.generate_instances_for_object_sync (comp, (time_t) data_range.first_dt.to_unix (), (time_t) data_range.last_dt.to_unix (), null, (comp, start, end) => {
-                var event = new ECal.Component.from_icalcomponent (comp);
+                client.generate_instances_for_object_sync (comp, (time_t) data_range.first_dt.to_unix (), (time_t) data_range.last_dt.to_unix (), null, (comp, start, end) => {
+                    var event = new ECal.Component.from_icalcomponent (comp);
 #else
-            client.generate_instances_for_object_sync (comp, (time_t) data_range.first_dt.to_unix (), (time_t) data_range.last_dt.to_unix (), (event, start, end) => {
+                client.generate_instances_for_object_sync (comp, (time_t) data_range.first_dt.to_unix (), (time_t) data_range.last_dt.to_unix (), (event, start, end) => {
 #endif
-                debug_event (source, event);
-                events_for_source.set (uid, event);
-                updated_events.add (event);
-                return true;
-            });
+                    event.set_data<E.Source>("source", source);
+                    debug_event (source, event, "MODIFIED - GENERATED");
+                    events_for_source.set (uid, event);
+                    added_events.add (event);
+                    return true;
+                });
+            } else {
+                debug_event (source, events_for_uid.to_array ()[0], "MODIFIED - UPDATED");
+                updated_events.add (events_for_uid.to_array ()[0]);
+            }
         });
 
+        events_removed (source, removed_events.read_only_view);
+        events_added (source, added_events.read_only_view);
         events_updated (source, updated_events.read_only_view);
     }
 
@@ -486,6 +501,7 @@ public class Maya.Model.CalendarModel : Object {
                 return;
 
             var comps = events.get (cid.get_uid ());
+            events.remove_all (cid.get_uid ());
             foreach (ECal.Component event in comps) {
                 removed_events.add (event);
                 debug_event (source, event);
